metrics and loss split
