+++ composer/algorithms/progressive_resizing/progressive_resizing.py
@@ -15,7 +15,7 @@ import torchvision.transforms.functional
 
 from composer.core import Algorithm, Event, State
 from composer.loggers import Logger
-from composer.models.loss import _check_for_index_targets
+from composer.loss.loss import _check_for_index_targets
 
 log = logging.getLogger(__name__)
