+++ composer/models/base.py
@@ -12,7 +12,8 @@ from torchmetrics import Metric, MetricCollection
 from torchmetrics.classification import Accuracy
 
 from composer.core.types import Batch, BatchPair
-from composer.models.loss import CrossEntropyLoss, soft_cross_entropy
+from composer.metrics.metrics import CrossEntropyLoss
+from composer.loss.loss import soft_cross_entropy
 
 __all__ = ["ComposerClassifier", "ComposerModel"]
